requirements.txt
uvicorn
sqlalchemy
fastapi
pydantic==1.7.4
requests
starlette==0.14.2
jinja2
pdoc3
pyinstrument
flask
minio
boto3
fastapi-login
kubernetes
python-dotenv
python-multipart
py-healthcheck
numpy==1.21.3
sklearn